Feat: 사용자 비교

+++ src/main/java/org/example/flowday/domain/course/course/controller/CourseController.java
@@ -4,9 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.example.flowday.domain.course.course.dto.CourseReqDTO;
 import org.example.flowday.domain.course.course.dto.CourseResDTO;
 import org.example.flowday.domain.course.course.dto.PageReqDTO;
+import org.example.flowday.domain.course.course.exception.CourseException;
+import org.example.flowday.domain.course.course.repository.CourseRepository;
 import org.example.flowday.domain.course.course.service.CourseService;
+import org.example.flowday.domain.member.entity.Member;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 public class CourseController {
 
     private final CourseService courseService;
+    private final CourseRepository courseRepository;
 
     // 코스 생성
     @PostMapping
@@ -36,7 +41,14 @@ public class CourseController {
 
     // 코스 삭제
     @DeleteMapping("/{course_id}")
-    public ResponseEntity<Void> deleteCourse(@PathVariable("course_id") Long courseId) {
+    public ResponseEntity<Void> deleteCourse(
+            @PathVariable("course_id") Long courseId,
+            @AuthenticationPrincipal Member user
+    ) {
+        if(!user.getId().equals(courseRepository.findById(courseId).get().getMember().getId())) {
+            throw CourseException.FORBIDDEN.get();
+        }
+
         courseService.removeCourse(courseId);
 
         return ResponseEntity.noContent().build();
